Skill Screen Tests Paramertrized
