+++ tests/test_2.py
@@ -1,64 +1,129 @@
 import pytest
-from pages.new_user_account import printNewAccountScreen
-from pages.main_menu import printMainMenu
 from pages.skill_search import printSkillScreen
 from pages.new_user_account import saveDatabase
 from tests.shared import JSONFP2
 import json
 
 
-def testCreateAccountUnder5(monkeypatch, capfd):
-    # Make sure Json is clear
-    # Test with 3 accounts
-    threeAccounts = [
-        {
-            "username": "dummy",
-            "password": "Password1!",
-            "firstname": "Jo",
-            "lastname": "Mama",
-            "connections": [],
-        },
-        {
-            "username": "sillyBoi",
-            "password": "Password2@",
-            "firstname": "Dee",
-            "lastname": "Snuts",
-            "connections": ["notKaren", "dummy"],
-        },
-        {
-            "username": "dummyDude",
-            "password": "Password2@",
-            "firstname": "Dee",
-            "lastname": "Snuts",
-            "connections": ["admin"],
-        },
-    ]
-
-    saveDatabase(JSONFP2, threeAccounts)
-
-    input_generator = iter(["usernew", "Johnathan", "Blow", "P@ssw0rd", "P@ssw0rd"])
-    # monkeypatch.setattr(builtins, "input", lambda : next(input_generator))
+@pytest.mark.parametrize(
+    "mock_input,responses,expectedReturn",
+    [
+        (
+            ["1", "anything"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                "*** Learn ",
+                "under construction, input anything to return",
+            ],
+            None,
+        ),
+        (
+            ["2", "anything"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                "*** Learn ",
+                "under construction, input anything to return",
+            ],
+            None,
+        ),
+        (
+            ["3", "anything"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                "*** Learn ",
+                "under construction, input anything to return",
+            ],
+            None,
+        ),
+        (
+            ["4", "anything"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                "*** Learn ",
+                "under construction, input anything to return",
+            ],
+            None,
+        ),
+        (
+            ["5", "anything"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                "*** Learn ",
+                "under construction, input anything to return",
+            ],
+            None,
+        ),
+        (
+            ["6"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+            ],
+            0,
+        ),
+        (
+            ["FoamEarplugs"],
+            [
+                "*** Learn a skill ***",
+                "1 - Learn how to skate",
+                "2 - Learn how to cook",
+                "3 - Learn how to drive",
+                "4 - Learn how to paint",
+                "5 - Learn how to whistle",
+                "6 - Return to main menu",
+                'Invalid selection please input "1" or "2" or "3" or "4" or "5" or "6"',
+            ],
+            0,
+        ),
+    ],
+    ids=["1-Skill", "2-Skill", "3-Skill", "4-Skill", "5-Skill", "6-Close", "InvalidSelection"],
+)
+def testSkillsScreen(mock_input, responses, expectedReturn, monkeypatch, capfd):
+    input_generator = iter(mock_input)
     monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
 
     try:
-        # captured = capfd.readouterr()
-        printNewAccountScreen()
+        assert printSkillScreen() == expectedReturn
     except StopIteration:
         pass
     captured = capfd.readouterr()
-
-    title = "*** Create a new user account ***\n"
-    firstname = "First name:"
-    lastname = "Last name:"
-    username = "Username:"
-    password = "Password:"
-
-    # Better approach than whole-sale output assertions
-    assert title in captured.out
-    assert firstname in captured.out
-    assert lastname in captured.out
-    assert username in captured.out
-    assert password in captured.out
+    for r in responses:
+        assert r in captured.out
 
     # Possibly delete below code, it does not test correctly
     # username = "Andrew"
@@ -82,24 +147,25 @@ def testCreateAccountUnder5(monkeypatch, capfd):
     # )
     # assert main_menu in captured.out
 
-@pytest.mark.parametrize("input_value", ["1", "2", "3", "4", "5"])
-def testSkillsUnderConstruction(input_value, monkeypatch, capfd):
-    input_generator = iter([input_value])
-    monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
 
-    try:
-        printSkillScreen()
-    except StopIteration:
-        pass
+# @pytest.mark.parametrize("input_value", ["1", "2", "3", "4", "5"])
+# def testSkillsUnderConstruction(input_value, monkeypatch, capfd):
+#     input_generator = iter([input_value])
+#     monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
 
-    captured = capfd.readouterr()
-    assert "under construction, input anything to return" in captured.out
+#     try:
+#         printSkillScreen()
+#     except StopIteration:
+#         pass
+
+#     captured = capfd.readouterr()
+#     assert "under construction, input anything to return" in captured.out
 
 
-def testSkillsReturnButton(monkeypatch, capfd):
-    input_generator = iter(["6"])  # causes skill screen to break
-    monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
-    assert printSkillScreen() == 0  # break from skill screen loop causes return 0
+# def testSkillsReturnButton(monkeypatch, capfd):
+#     input_generator = iter(["6"])  # causes skill screen to break
+#     monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
+#     assert printSkillScreen() == 0  # break from skill screen loop causes return 0
 
     # monkeypatch.setattr("builtins.input", lambda: next(input_generator))
     # try:
@@ -121,15 +187,15 @@ def testSkillsReturnButton(monkeypatch, capfd):
     # assert skill in captured.out
 
 
-@pytest.mark.parametrize("input_value", ["9", "0", "[1", "d", "r"])
-def testSkillsWrongInput(input_value, monkeypatch, capfd):
-    input_generator = iter([input_value])
-    monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
+# @pytest.mark.parametrize("input_value", ["9", "0", "[1", "d", "r"])
+# def testSkillsWrongInput(input_value, monkeypatch, capfd):
+#     input_generator = iter([input_value])
+#     monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
 
-    try:
-        printSkillScreen()
-    except StopIteration:
-        pass
+#     try:
+#         printSkillScreen()
+#     except StopIteration:
+#         pass
 
-    captured = capfd.readouterr()
-    assert 'Invalid selection please input "1" or "2" or "3" or "4" or "5" or "6"' in captured.out
+#     captured = capfd.readouterr()
+#     assert 'Invalid selection please input "1" or "2" or "3" or "4" or "5" or "6"' in captured.out
