+++ tests/test_initial_screen.py
@@ -1,28 +1,6 @@
 from main import printInitialScreen
 from pages.initial_screen import printInitialScreen
 import pytest
-
-
-def testPrintInitialScreen(monkeypatch, capfd):
-    input_generator = iter(["2"])
-    monkeypatch.setattr("builtins.input", lambda _: next(input_generator))
-
-    try:
-        printInitialScreen()
-    except StopIteration:
-        pass
-
-    captured = capfd.readouterr()
-
-    title = "*** Welcome to InCollege ***"
-    existingUser = "1 - Login as existing user"
-    newAccount = "2 - Create a new InCollege account"
-
-    assert title in captured.out
-    assert existingUser in captured.out
-    assert newAccount in captured.out
-
-
 @pytest.mark.parametrize(
     "mock_input,responses,expectedReturn",
     [
@@ -41,7 +19,7 @@ def testPrintInitialScreen(monkeypatch, capfd):
             [],
         ),
         (
-            ["anything", "2"], # New account screen
+            ["anything", "2"],  # New account screen
             [
                 "*** Welcome to InCollege ***",
                 "Here is a story from one of our users:",
@@ -53,7 +31,7 @@ def testPrintInitialScreen(monkeypatch, capfd):
             [],
         ),
         (
-            ["anything", "3"], #Search for friend user
+            ["anything", "3"],  # Search for friend user
             [
                 "*** Welcome to InCollege ***",
                 "Here is a story from one of our users:",
@@ -92,6 +70,7 @@ def testPrintInitialScreen(monkeypatch, capfd):
             [],
         ),
     ],
+    ids=["1-Login", "2-CreateNewAccount", "3-FindSomeone", "X-Close", "InvalidSelection"],
 )
 def testPrintInitial(mock_input, responses, expectedReturn, monkeypatch, capfd):
     input_generator = iter(mock_input)
